requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
Js2Py==0.71
lorem==0.1.1
pyjsparser==2.7.1
python-dotenv==0.20.0
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
six==1.16.0
smmap==5.0.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
